Update 0.1.1-beta
In this version of the code I solved the word analysis problem that came up when the iteration variable o word was not found in the dictionary keys but in the values. This version of the code can analyze simple describing sentences about a  person and store them in a foreign file. It can also detect if someone writes that for instance "I'm not boring", in this case it will browse for the opposite of boring, then store the opposite of boring in the file that was created in order to store fundamental informations about the user.