Create Points_to_line_converter
This math program converts two given points into the equation of a line, which passes through them.
After starting, it asks the User for the coordinates of the two points. Using these coordinates, it calculates a direction vector, then a normal vector. At the end, using the normal vector and the second point, it calculates the equation of the line, which passes through both points.